refactor: Lotto 클래스 생성자 private 변경 및 정적 팩토리 메서드 구현
